feat(grasshopper): add artefact-first Load and Publish components (#1512)
* feat(grasshopper): add artefact-first Load and Publish components

* refactor(grasshopper): drop the Source output from the new Load components
