+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Components/Operations/Send/SendAsyncComponentBase.cs
@@ -1,0 +1,686 @@
+using System.Diagnostics;
+using System.Timers;
+using Grasshopper.GUI;
+using Grasshopper.GUI.Canvas;
+using Grasshopper.Kernel;
+using Grasshopper.Kernel.Attributes;
+using Grasshopper.Kernel.Types;
+using GrasshopperAsyncComponent;
+using Microsoft.Extensions.DependencyInjection;
+using Rhino;
+using Speckle.Connectors.Common.Analytics;
+using Speckle.Connectors.Common.Operations;
+using Speckle.Connectors.GrasshopperShared.HostApp;
+using Speckle.Connectors.GrasshopperShared.Parameters;
+using Speckle.Connectors.GrasshopperShared.Registration;
+using Speckle.Sdk;
+using Speckle.Sdk.Api;
+using Speckle.Sdk.Common;
+using Speckle.Sdk.Credentials;
+using Speckle.Sdk.Models.Extensions;
+using Speckle.Sdk.Pipelines.Progress;
+
+namespace Speckle.Connectors.GrasshopperShared.Components.Operations.Send;
+
+/// <summary>
+/// Shared machinery for the asynchronous Publish components. Subclasses differ only in whether they write a 4.0
+/// bundle (<see cref="UseArtifacts"/>) and whether they take model-wide properties
+/// (<see cref="HasModelPropertiesInput"/>).
+/// </summary>
+public abstract class SendAsyncComponentBase : GH_AsyncComponent<SendAsyncComponentBase>
+{
+  protected SendAsyncComponentBase(
+    string name,
+    string nickname,
+    string description,
+    string category,
+    string subCategory
+  )
+    : base(name, nickname, description, category, subCategory)
+  {
+    BaseWorker = new SendComponentWorker(this);
+  }
+
+  public override void CreateAttributes() => m_attributes = new SendAsyncComponentAttributes(this);
+
+  /// <summary>False falls through to the v3 routes - see <c>SendOperation.Send</c>'s useArtifacts parameter.</summary>
+  public abstract bool UseArtifacts { get; }
+
+  /// <summary>Whether this variant exposes the model-wide properties input. It is the last input, so no other index moves.</summary>
+  protected abstract bool HasModelPropertiesInput { get; }
+
+  /// <summary>Hook for the deprecated variant to say what it created. No-op otherwise.</summary>
+  public virtual void OnPublished() { }
+
+  public ComponentState CurrentComponentState { get; set; } = ComponentState.NeedsInput;
+  public bool AutoSend { get; set; }
+  public bool JustPastedIn { get; set; }
+  public double OverallProgress { get; set; }
+  public string? Url { get; set; }
+  public IClient ApiClient { get; set; }
+  public SpeckleUrlModelResource? UrlModelResource { get; set; }
+  public SpeckleCollectionWrapperGoo? RootCollectionWrapper { get; set; }
+  public SpecklePropertyGroupGoo? RootProperties { get; private set; }
+  public SpeckleUrlModelResource? OutputParam { get; set; }
+  public bool HasMultipleInputs { get; set; }
+
+  public string? VersionMessage { get; private set; }
+
+  protected override void RegisterInputParams(GH_InputParamManager pManager)
+  {
+    // speckle model
+    pManager.AddParameter(new SpeckleUrlModelResourceParam());
+
+    // collection / data
+    pManager.AddGenericParameter(
+      "Collection",
+      "collection",
+      "The collections, data objects, or geometries to publish",
+      GH_ParamAccess.list
+    );
+
+    // version message
+    pManager.AddTextParameter("Version Message", "versionMessage", "The version message", GH_ParamAccess.item);
+    pManager[2].Optional = true;
+
+    if (HasModelPropertiesInput)
+    {
+      // model-wide props (see cnx-2722)
+      pManager.AddParameter(
+        new SpecklePropertyGroupParam(),
+        "Properties",
+        "properties",
+        "Optional model-wide properties to attach to the root collection",
+        GH_ParamAccess.item
+      );
+      pManager[3].Optional = true;
+    }
+  }
+
+  protected override void RegisterOutputParams(GH_OutputParamManager pManager)
+  {
+    pManager.AddParameter(new SpeckleUrlModelResourceParam());
+    pManager.AddTextParameter("Version ID", "V", "ID of the created version", GH_ParamAccess.item);
+  }
+
+  public override void AppendAdditionalMenuItems(ToolStripDropDown menu)
+  {
+    static void Open(string url)
+    {
+      var psi = new ProcessStartInfo { FileName = url, UseShellExecute = true };
+      Process.Start(psi);
+    }
+
+    base.AppendAdditionalMenuItems(menu);
+
+    Menu_AppendSeparator(menu);
+
+    var autoSendMi = Menu_AppendItem(
+      menu,
+      "Publish automatically",
+      (_, _) =>
+      {
+        AutoSend = !AutoSend;
+        RhinoApp.InvokeOnUiThread(
+          (Action)
+            delegate
+            {
+              OnDisplayExpired(true);
+            }
+        );
+      },
+      true,
+      AutoSend
+    );
+    autoSendMi.ToolTipText =
+      "Toggle automatic data publishing. If set, any change in any of the input parameters of this component will start publishing.\n Please be aware that if a new publish starts before an old one is finished, the previous operation is cancelled.";
+
+    if (Url != null)
+    {
+      Menu_AppendSeparator(menu);
+
+      Menu_AppendItem(menu, $"View created model online ↗", (_, _) => Open(Url));
+    }
+
+    Menu_AppendSeparator(menu);
+
+    if (CurrentComponentState == ComponentState.Sending)
+    {
+      Menu_AppendItem(
+        menu,
+        "Cancel Publish",
+        (_, _) =>
+        {
+          CurrentComponentState = ComponentState.Expired;
+          RequestCancellation();
+        }
+      );
+    }
+  }
+
+  protected override void SolveInstance(IGH_DataAccess da)
+  {
+    var multipleResources = Params.Input[0].VolatileData.HasInputCountGreaterThan(1);
+
+    HasMultipleInputs = multipleResources;
+
+    if (HasMultipleInputs)
+    {
+      AddRuntimeMessage(
+        GH_RuntimeMessageLevel.Error,
+        "Only one single model can be published to from this node. To send to multiple models, please use different publish components."
+      );
+      return;
+    }
+
+    using var scope = PriorityLoader.CreateScopeForActiveDocument();
+
+    // We need to call this always in here to be able to react and set events :/
+    if (!ParseInput(da, scope))
+    {
+      return;
+    }
+
+    if (
+      (AutoSend || CurrentComponentState == ComponentState.Ready || CurrentComponentState == ComponentState.Sending)
+      && !JustPastedIn
+    )
+    {
+      CurrentComponentState = ComponentState.Sending;
+
+      // Delegate control to parent async component.
+      base.SolveInstance(da);
+      return;
+    }
+
+    if (JustPastedIn)
+    {
+      // Set output data in a "first run" event. Note: we are not persisting the actual "sent" object as it can be very big.
+      base.SolveInstance(da);
+    }
+    else
+    {
+      da.SetData(0, OutputParam);
+      CurrentComponentState = ComponentState.Expired;
+      Message = "Expired";
+      OnDisplayExpired(true);
+    }
+  }
+
+  public override void RemovedFromDocument(GH_Document document)
+  {
+    RequestCancellation();
+    base.RemovedFromDocument(document);
+  }
+
+  public override void DisplayProgress(object sender, ElapsedEventArgs e)
+  {
+    if (WorkerCount == 0)
+    {
+      return;
+    }
+
+    Message = "";
+    var total = 0.0;
+    foreach (var kvp in ProgressReports)
+    {
+      Message += $"{kvp.Key}: {kvp.Value}\n";
+      total += kvp.Value;
+    }
+
+    OverallProgress = total / ProgressReports.Keys.Count;
+
+    RhinoApp.InvokeOnUiThread(
+      (Action)
+        delegate
+        {
+          OnDisplayExpired(true);
+        }
+    );
+  }
+
+  public override void DocumentContextChanged(GH_Document document, GH_DocumentContext context)
+  {
+    switch (context)
+    {
+      case GH_DocumentContext.Loaded:
+        OnDisplayExpired(true);
+        break;
+
+      case GH_DocumentContext.Unloaded:
+        // Will execute every time a document becomes inactive (in background or closing file.)
+        //Correctly dispose of the client when changing documents to prevent subscription handlers being called in background.
+        RequestCancellation();
+        break;
+    }
+
+    base.DocumentContextChanged(document, context);
+  }
+
+  /// <returns>
+  /// <c>false</c> if any validation failed and the send should be aborted; <c>true</c> if inputs are valid.
+  /// </returns>
+  private bool ParseInput(IGH_DataAccess da, IServiceScope scope)
+  {
+    SpeckleUrlModelResource? dataInput = null;
+    da.GetData(0, ref dataInput);
+    if (dataInput is null)
+    {
+      UrlModelResource = null;
+      TriggerAutoSave();
+      return false;
+    }
+
+    UrlModelResource = dataInput;
+    try
+    {
+      Account? account = dataInput.Account.GetAccount(scope);
+      if (account is null)
+      {
+        throw new SpeckleAccountManagerException($"No default account was found");
+      }
+
+      ApiClient?.Dispose();
+      ApiClient = scope.Get<IClientFactory>().Create(account);
+    }
+    catch (Exception e) when (!e.IsFatal())
+    {
+      AddRuntimeMessage(GH_RuntimeMessageLevel.Error, e.ToFormattedString());
+    }
+
+    List<IGH_Goo> inputGoos = new();
+    da.GetDataList(1, inputGoos);
+
+    if (inputGoos.Count == 0)
+    {
+      RootCollectionWrapper = null;
+      TriggerAutoSave();
+      return false;
+    }
+
+    SpeckleCollectionWrapper? rootBase;
+
+    // filter out nulls just to check if we can use the fast path
+    var nonNullGoos = inputGoos.Where(x => x != null).ToList();
+
+    // fast path: if there's exactly one valid item and it's a collection, use it directly
+    if (nonNullGoos.Count == 1 && nonNullGoos[0] is SpeckleCollectionWrapperGoo singleCollection)
+    {
+      rootBase = singleCollection.Value.DeepCopy();
+    }
+    else
+    {
+      // mixed inputs: construct a root collection using the document name (CNX-3175)
+      var docName = SendComponentBase.GetGrasshopperFileInfo().fileName ?? "Unnamed Document";
+      if (
+        docName.EndsWith(".gh", StringComparison.OrdinalIgnoreCase)
+        || docName.EndsWith(".ghx", StringComparison.OrdinalIgnoreCase)
+      )
+      {
+        docName = Path.GetFileNameWithoutExtension(docName);
+      }
+
+      rootBase = new SpeckleCollectionWrapper
+      {
+        Base = new Speckle.Sdk.Models.Collections.Collection() { name = docName },
+        Path = [docName],
+        Color = null,
+        Material = null,
+        Name = docName,
+      };
+
+      int skippedCount = 0;
+      foreach (var obj in inputGoos)
+      {
+        if (obj is SpeckleCollectionWrapperGoo collectionGoo)
+        {
+          var colClone = (SpeckleCollectionWrapperGoo)collectionGoo.Duplicate();
+          colClone.Value.Path = rootBase.Path;
+          rootBase.Elements.AddRange(colClone.Value.Elements);
+        }
+        else if (obj is SpeckleDataObjectWrapperGoo dataObjectWrapperGoo)
+        {
+          var dataObjectWrapper = dataObjectWrapperGoo.Value.DeepCopy();
+          dataObjectWrapper.Path = rootBase.Path;
+          dataObjectWrapper.Parent = rootBase;
+          rootBase.Elements.Add(dataObjectWrapper);
+        }
+        // reject bare geometry — collections may only contain Data Objects or sub-collections.
+        // SpeckleGeometry is easily wrapped: wire your geometry through a 'Speckle Data Object' component first.
+        else if (obj?.ToSpeckleGeometryWrapper() is not null)
+        {
+          AddRuntimeMessage(
+            GH_RuntimeMessageLevel.Error,
+            "Speckle Geometry cannot be added directly to a Collection. "
+              + "Use a 'Speckle Data Object' component to wrap your geometry first, then pipe it into the Collection."
+          );
+          return false;
+        }
+        else
+        {
+          rootBase.Elements.Add(null);
+          skippedCount++;
+        }
+      }
+
+      if (skippedCount > 0)
+      {
+        AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, $"Skipped {skippedCount} unsupported object(s).");
+      }
+    }
+
+    RootCollectionWrapper = new SpeckleCollectionWrapperGoo(rootBase);
+
+    string? versionMessage = null;
+    da.GetData(2, ref versionMessage);
+    VersionMessage = versionMessage;
+
+    SpecklePropertyGroupGoo? rootPropsGoo = null;
+    if (HasModelPropertiesInput)
+    {
+      da.GetData(3, ref rootPropsGoo);
+
+      // validate single properties group
+      if (Params.Input[3].VolatileData.DataCount > 1)
+      {
+        AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "Only one Model Properties group is allowed");
+        return false;
+      }
+    }
+
+    RootProperties = rootPropsGoo;
+    return true;
+  }
+}
+
+public class SendComponentWorker : WorkerInstance<SendAsyncComponentBase>
+{
+  public SendComponentWorker(
+    SendAsyncComponentBase p,
+    string id = "baseWorker",
+    CancellationToken cancellationToken = default
+  )
+    : base(p, id, cancellationToken) { }
+
+  private Stopwatch? _stopwatch;
+  public SpeckleUrlModelResource? OutputParam { get; set; }
+  public string? OutputVersionId { get; set; }
+  private List<(GH_RuntimeMessageLevel, string)> RuntimeMessages { get; } = new();
+
+  public override WorkerInstance<SendAsyncComponentBase> Duplicate(string id, CancellationToken cancellationToken)
+  {
+    return new SendComponentWorker(Parent, id, cancellationToken);
+  }
+
+  public override void GetData(IGH_DataAccess da, GH_ComponentParamServer p)
+  {
+    _stopwatch = new Stopwatch();
+    _stopwatch.Start();
+    OutputVersionId = null;
+  }
+
+  public override void SetData(IGH_DataAccess da)
+  {
+    _stopwatch.NotNull("GetData must be called before SetData").Stop();
+
+    if (Parent.JustPastedIn)
+    {
+      Parent.JustPastedIn = false;
+      da.SetData(0, Parent.OutputParam);
+      da.SetData(1, OutputVersionId);
+      return;
+    }
+
+    if (CancellationToken.IsCancellationRequested)
+    {
+      Parent.CurrentComponentState = ComponentState.Expired;
+      return;
+    }
+
+    foreach (var (level, message) in RuntimeMessages)
+    {
+      Parent.AddRuntimeMessage(level, message);
+    }
+
+    da.SetData(0, OutputParam);
+    da.SetData(1, OutputVersionId);
+
+    // not added to RuntimeMessages - the success remarks below are gated on that list being empty
+    if (OutputVersionId != null)
+    {
+      Parent.OnPublished();
+    }
+
+    Parent.CurrentComponentState = ComponentState.UpToDate;
+    Parent.OutputParam = OutputParam; // ref the outputs in the parent too, so we can serialise them on write/read
+    Parent.OverallProgress = 0;
+
+    var hasWarnings = RuntimeMessages.Count > 0;
+    if (!hasWarnings)
+    {
+      /* POC: cannot use GetTotalChildrenCount() on the root collection, because this contains subcollection wrappers which are not recognized by our typeloader. Will throw exception as result.
+      Parent.AddRuntimeMessage(
+        GH_RuntimeMessageLevel.Remark,
+        $"Successfully published {((SendAsyncComponentBase)Parent).RootCollectionWrapper?.Value.Collection.GetTotalChildrenCount()} objects to Speckle."
+      );
+      */
+      Parent.AddRuntimeMessage(
+        GH_RuntimeMessageLevel.Remark,
+        "Successfully published to Speckle. Right-click on the component to view online."
+      );
+      Parent.AddRuntimeMessage(
+        GH_RuntimeMessageLevel.Remark,
+        $"Publish duration: {_stopwatch.ElapsedMilliseconds / 1000f}s"
+      );
+    }
+  }
+
+  public override async Task DoWork(Action<string, double> reportProgress, Action done)
+  {
+    if (Parent.JustPastedIn || Parent.HasMultipleInputs)
+    {
+      return;
+    }
+
+    try
+    {
+      await Send(reportProgress);
+      done();
+    }
+    catch (OperationCanceledException) when (CancellationToken.IsCancellationRequested)
+    {
+      RuntimeMessages.Add((GH_RuntimeMessageLevel.Remark, "Operation cancelled"));
+      Parent.CurrentComponentState = ComponentState.Expired;
+      //No need to call `done()` - GrasshopperAsyncComponent assumes immediate cancel,
+      //thus it has already performed clean-up actions that would normally be done on `done()`
+    }
+    catch (Exception ex) when (!ex.IsFatal())
+    {
+      RuntimeMessages.Add((GH_RuntimeMessageLevel.Error, ex.ToFormattedString()));
+      done();
+    }
+  }
+
+  private async Task Send(Action<string, double> reportProgress)
+  {
+    SpeckleUrlModelResource? urlModelResource = Parent.UrlModelResource;
+    if (urlModelResource is null)
+    {
+      throw new InvalidOperationException("Url Resource was null");
+    }
+
+    SpeckleCollectionWrapperGoo? rootCollectionWrapper = Parent.RootCollectionWrapper;
+    if (rootCollectionWrapper is null)
+    {
+      throw new InvalidOperationException("Root Collection was null");
+    }
+
+    // safe to always create new wrapper since users cannot create SpeckleRootCollectionWrapper directly - it's only
+    // constructed here from the Collection + Model Properties inputs.
+    // if this changes, then we need to update below!
+    var rootWrapper = new SpeckleRootCollectionWrapper(rootCollectionWrapper.Value, Parent.RootProperties?.Unwrap());
+
+    rootCollectionWrapper = new SpeckleRootCollectionWrapperGoo(rootWrapper);
+
+    // Step 1 - SEND TO SERVER
+    var sendInfo = await urlModelResource.GetSendInfo(Parent.ApiClient, CancellationToken).ConfigureAwait(false);
+
+    var (fileName, fileBytes) = SendComponentBase.GetGrasshopperFileInfo();
+    var progress = new Progress<CardProgress>(p =>
+    {
+      reportProgress(Id, p.Progress ?? 0);
+      //sendComponent.Message = $"{p.Status}";
+    });
+    using var scope = PriorityLoader.CreateScopeForActiveDocument();
+    var sendOperation = scope.ServiceProvider.GetRequiredService<SendOperation<SpeckleCollectionWrapperGoo>>();
+    (SendOperationResult result, string versionId, string? ingestionId) = await sendOperation
+      .Send(
+        [rootCollectionWrapper],
+        sendInfo,
+        fileName,
+        fileBytes,
+        Parent.VersionMessage,
+        progress,
+        true,
+        CancellationToken,
+        Parent.UseArtifacts
+      )
+      .ConfigureAwait(false);
+
+    if (ingestionId != null)
+    {
+      Parent.Message = "Remote processing";
+      var ingestionTracker = scope.ServiceProvider.GetRequiredService<IngestionTracker>();
+      versionId = await ingestionTracker
+        .WaitForIngestionCompletion(
+          Parent.ApiClient,
+          sendInfo.ProjectId,
+          ingestionId,
+          reportProgress,
+          Id,
+          CancellationToken
+        )
+        .ConfigureAwait(false);
+    }
+
+    // TODO: If we have NodeRun events later, better to have `ComponentTracker` to use across components
+    var customProperties = new Dictionary<string, object>() { { "isAsync", true }, { "auto", Parent.AutoSend } };
+    if (sendInfo.WorkspaceId != null)
+    {
+      customProperties.Add("workspace_id", sendInfo.WorkspaceId);
+    }
+
+    var mixPanelManager = scope.ServiceProvider.GetRequiredService<IMixPanelManager>();
+    await mixPanelManager
+      .TrackEvent(MixPanelEvents.Send, Parent.ApiClient.Account, customProperties)
+      .ConfigureAwait(false);
+
+    SpeckleUrlModelVersionResource createdVersion = new(
+      new(sendInfo.Account.id, null, sendInfo.Account.serverInfo.url),
+      sendInfo.WorkspaceId,
+      sendInfo.ProjectId,
+      sendInfo.ModelId,
+      versionId
+    );
+    OutputParam = createdVersion;
+    OutputVersionId = versionId;
+    Parent.Url = $"{createdVersion.Account.Server}/projects/{sendInfo.ProjectId}/models/{sendInfo.ModelId}";
+  }
+}
+
+public class SendAsyncComponentAttributes : GH_ComponentAttributes
+{
+  public SendAsyncComponentAttributes(GH_Component owner)
+    : base(owner) { }
+
+  private Rectangle ButtonBounds { get; set; }
+
+  protected override void Layout()
+  {
+    base.Layout();
+
+    var baseRec = GH_Convert.ToRectangle(Bounds);
+    baseRec.Height += 26;
+
+    var btnRec = baseRec;
+    btnRec.Y = btnRec.Bottom - 26;
+    btnRec.Height = 26;
+    btnRec.Inflate(-2, -2);
+
+    Bounds = baseRec;
+    ButtonBounds = btnRec;
+  }
+
+  protected override void Render(GH_Canvas canvas, Graphics graphics, GH_CanvasChannel channel)
+  {
+    base.Render(canvas, graphics, channel);
+
+    var state = ((SendAsyncComponentBase)Owner).CurrentComponentState;
+
+    if (channel == GH_CanvasChannel.Objects)
+    {
+      if (((SendAsyncComponentBase)Owner).AutoSend)
+      {
+        using var autoSendButton = GH_Capsule.CreateTextCapsule(
+          ButtonBounds,
+          ButtonBounds,
+          GH_Palette.Blue,
+          "Auto Publish",
+          2,
+          0
+        );
+
+        autoSendButton.Render(graphics, Selected, Owner.Locked, false);
+      }
+      else
+      {
+        var palette =
+          state == ComponentState.Expired || state == ComponentState.UpToDate
+            ? GH_Palette.Black
+            : GH_Palette.Transparent;
+
+        var text = state == ComponentState.Sending ? "Publishing..." : "Publish";
+
+        using var button = GH_Capsule.CreateTextCapsule(
+          ButtonBounds,
+          ButtonBounds,
+          palette,
+          text,
+          2,
+          state == ComponentState.Expired ? 10 : 0
+        );
+        button.Render(graphics, Selected, Owner.Locked, false);
+      }
+    }
+  }
+
+  public override GH_ObjectResponse RespondToMouseDown(GH_Canvas sender, GH_CanvasMouseEvent e)
+  {
+    if (e.Button == MouseButtons.Left)
+    {
+      if (((RectangleF)ButtonBounds).Contains(e.CanvasLocation))
+      {
+        // Ignore button presses on a component with invalid inputs to prevent multiple sends
+        if ((Owner as SendAsyncComponentBase)?.HasMultipleInputs == true)
+        {
+          return GH_ObjectResponse.Handled;
+        }
+        if (((SendAsyncComponentBase)Owner).AutoSend)
+        {
+          ((SendAsyncComponentBase)Owner).AutoSend = false;
+          Owner.OnDisplayExpired(true);
+          return GH_ObjectResponse.Handled;
+        }
+        if (((SendAsyncComponentBase)Owner).CurrentComponentState == ComponentState.Sending)
+        {
+          return GH_ObjectResponse.Handled;
+        }
+        ((SendAsyncComponentBase)Owner).CurrentComponentState = ComponentState.Ready;
+        Owner.ExpireSolution(true);
+        return GH_ObjectResponse.Handled;
+      }
+    }
+
+    return base.RespondToMouseDown(sender, e);
+  }
+}
