+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Speckle.Connectors.GrasshopperShared.projitems
@@ -55,6 +55,10 @@
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\GrasshopperColorUnpacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\GrasshopperArtefactObjectBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\GrasshopperCollectionRebuilder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Receive\ReceiveComponentBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Receive\ReceiveArtefactComponent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Receive\ReceiveAsyncComponentBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Receive\ReceiveArtefactAsyncComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Receive\ReceiveAsyncComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Receive\ReceiveComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Send\GrasshopperArtifactRootObjectBuilder.cs" />
@@ -63,6 +67,10 @@
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Send\GrasshopperContinuousTraversalBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Send\GrasshopperColorPacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\IngestionTracker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\SendComponentBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\SendArtefactComponent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\SendAsyncComponentBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\SendArtefactAsyncComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\SendAsyncComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Send\SendComponent.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\SpeckleSelectModelComponent.cs" />
